Fix issue #1 Ios platform ready
The problem is that this plugin has a ready event name conflict with
cordova-plugin-device.  If this plugin and cordova-plugin-device are
added to the same hybrid app, the deviceready event is never fired.
Since Ionic app scaffolding adds cordova-plugin-device by default, the
problem appears in Ionic app more easily.

The fix is to use a ready event name more specific to this plugin.